refactor hir optimization pass
