+++ csrc/host_ir/pass/insert_deallocations.h
@@ -8,11 +8,20 @@
 #pragma once
 
 #include <host_ir/container.h>
+#include <host_ir/pass/optimization_pass.h>
 
-namespace nvfuser::hir {
+namespace nvfuser::hir_pass {
 
 /* For each input in every expression in the container, find the index of its
  * last use and insert a deallocate directly after */
-void insertDeallocations(HostIrContainer* hic);
+class InsertDeallocations : public OptimizationPass<InsertDeallocations> {
+  friend class OptimizationPass<InsertDeallocations>;
 
-} // namespace nvfuser::hir
+ protected:
+  void passImplementation(Fusion* fusion) override;
+  static constexpr std::string_view name() {
+    return "InsertDeallocations";
+  }
+};
+
+} // namespace nvfuser::hir_pass
