+++ csrc/host_ir/pass/stream_parallel_type.cpp
@@ -18,7 +18,7 @@
 #include <ops/all_ops.h>
 #include <ops/utils.h>
 
-namespace nvfuser::hir {
+namespace nvfuser::hir_pass {
 
 namespace {
 
@@ -399,7 +399,7 @@ std::vector<Expr*> addStreamManagement(std::vector<Expr*> top_level_exprs) {
 // linear structure of the HostIrContainer::topLevelExpr to greedily merge the
 // adjacent compatible stream for-loop bodies. Ideally we should look at the dag
 // and use the segmenter.
-void StreamParallelType::runPass(Fusion* fusion) {
+void StreamParallelType::passImplementation(Fusion* fusion) {
   // Verify that input tensors don't have stream axes
   NVF_CHECK(
       std::all_of(
@@ -437,4 +437,4 @@ void StreamParallelType::runPass(Fusion* fusion) {
   hic->resetTopLevelExprs(top_level_exprs);
 }
 
-} // namespace nvfuser::hir
+} // namespace nvfuser::hir_pass
