+++ csrc/runtime/fusion_kernel_runtime.cpp
@@ -11,6 +11,7 @@
 #include <fusion_profiler.h>
 #include <fusion_segmenter.h>
 #include <host_ir/lower.h>
+#include <host_ir/pass/convert_op_to_communication.h>
 #include <host_ir/pass/insert_deallocations.h>
 #include <instrumentation.h>
 #include <ir/base_nodes.h>
@@ -519,9 +520,11 @@ void FusionKernelRuntime::compileFusionParallel(KernelArgumentHolder args) {
           NVF_ERROR(
               group_to_run->exprs().size() == 1,
               "Communication segments must contain only one Expr");
-          HostIrLower lower;
-          for (auto* expr : lower.lower(
-                   ir_cloner.clone(group_to_run->exprs().at(0)), deviceid)) {
+          for (auto* expr : hir_pass::ConvertOpToCommunication::
+                   ConvertSingleOpToCommunication(
+                       ir_cloner.clone(group_to_run->exprs().at(0)),
+                       deviceid,
+                       HostIrLowerParams())) {
             NVF_ERROR(
                 expr->isA<Communication>(),
                 "Exprs in a Communication group should be Communication");
@@ -554,7 +557,7 @@ void FusionKernelRuntime::compileFusionParallel(KernelArgumentHolder args) {
       hic->addOutput(ir_cloner.clone(out));
     }
 
-    insertDeallocations(hic.get());
+    hir_pass::InsertDeallocations().runPass(hic.get());
 
     hie_ = std::make_unique<hir::HostIrEvaluator>(
         std::move(hic), &Communicator::getInstance());
