+++ python/python_frontend/fusion_definition.cpp
@@ -455,7 +455,7 @@ std::pair<KernelArgumentHolder, std::vector<Sharding>> FusionDefinition::
         params.lower.communicator_backend = backend_type_;
         // Disable StreamParallelType pass temporarily as proper stream lowering
         // gets implemented
-        preseg_passes::OptimizationPassGuard<hir::StreamParallelType> guard(
+        hir_pass::OptimizationPassGuard<hir_pass::StreamParallelType> guard(
             false);
         scheds->multi_device_executor = std::make_unique<MultiDeviceExecutor>(
             std::make_unique<Fusion>(*scheds->preschedFusion()),
