+++ tests/cpp/test_host_ir_stream_lowering.cpp
@@ -35,8 +35,7 @@ TEST_F(HirLowerStreamTest, InputsAreNotStreamParallelized) {
   hic->addInput(tv);
   tv->axis(0)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 TEST_F(HirLowerStreamTest, Split) {
@@ -50,8 +49,7 @@ TEST_F(HirLowerStreamTest, Split) {
   tv1->split(0, 2);
   tv1->axis(0)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 TEST_F(HirLowerStreamTest, Merge) {
@@ -65,8 +63,7 @@ TEST_F(HirLowerStreamTest, Merge) {
   tv1->merge(0, 1);
   tv1->axis(0)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 TEST_F(HirLowerStreamTest, SingleSetOp) {
@@ -81,7 +78,7 @@ TEST_F(HirLowerStreamTest, SingleSetOp) {
   tv1->setMemoryType(MemoryType::Global);
   tv1->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -110,7 +107,7 @@ TEST_F(HirLowerStreamTest, SingleSetOpNonOutermost) {
   tv1->setMemoryType(MemoryType::Global);
   tv1->axis(1)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -142,7 +139,7 @@ TEST_F(HirLowerStreamTest, SingleBinaryOp) {
   tv2->setMemoryType(MemoryType::Global);
   tv2->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -153,7 +150,6 @@ TEST_F(HirLowerStreamTest, SingleBinaryOp) {
   auto options = at::TensorOptions().device(at::kCUDA, 0);
   at::Tensor tv0_input = at::rand({4, 4}, options);
   at::Tensor tv1_input = at::rand({4, 4}, options);
-  // std::unordered_map<Val*, PolymorphicValue> inputs = {{tv0, input}};
   auto output = hie.runWithInput({{tv0, tv0_input}, {tv1, tv1_input}})[0]
                     .as<at::Tensor>();
   auto expected_output = tv0_input + tv1_input;
@@ -177,7 +173,7 @@ TEST_F(HirLowerStreamTest, TwoSetOps) {
   tv1->axis(0)->parallelize(ParallelType::Stream);
   tv2->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 3);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -214,7 +210,7 @@ TEST_F(HirLowerStreamTest, ThreeSetOpsWithDisjointsForLoops) {
   tv1->axis(0)->parallelize(ParallelType::Stream);
   tv3->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 5);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -246,8 +242,7 @@ TEST_F(HirLowerStreamTest, ReductionUnsupported) {
   tv1->setMemoryType(MemoryType::Global);
   tv1->axis(0)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 TEST_F(HirLowerStreamTest, Reduction) {
@@ -262,7 +257,7 @@ TEST_F(HirLowerStreamTest, Reduction) {
   tv1->setMemoryType(MemoryType::Global);
   tv1->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -295,7 +290,7 @@ TEST_F(HirLowerStreamTest, Matmul_M) {
   c->setMemoryType(MemoryType::Global);
   c->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -331,7 +326,7 @@ TEST_F(HirLowerStreamTest, BatchedMatmul) {
   c->setMemoryType(MemoryType::Global);
   c->axis(0)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -367,7 +362,7 @@ TEST_F(HirLowerStreamTest, Matmul_N) {
   c->setMemoryType(MemoryType::Global);
   c->axis(1)->parallelize(ParallelType::Stream);
 
-  preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get());
+  hir_pass::StreamParallelType().runPass(hic.get());
 
   EXPECT_EQ(hic->topLevelExprs().size(), 2);
   EXPECT_TRUE(hic->topLevelExprs().at(0)->isA<kir::Allocate>());
@@ -403,8 +398,7 @@ TEST_F(HirLowerStreamTest, Matmul_K) {
   c->setMemoryType(MemoryType::Global);
   c->axis(-1)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 // We don's support PostOnStream because it does not support well pre-allocated
@@ -451,8 +445,7 @@ TEST_F(HirLowerStreamTest, DoNotSupportPostOnStream) {
 
   output->axis(-1)->parallelize(ParallelType::Stream);
 
-  EXPECT_ANY_THROW(
-      preseg_passes::OptimizationPass<StreamParallelType>::runPass(hic.get()));
+  EXPECT_ANY_THROW(hir_pass::StreamParallelType().runPass(hic.get()));
 }
 
 } // namespace hir
