+++ tests/cpp/test_multidevice_host_ir.cpp
@@ -365,7 +365,7 @@ using OverlapDistributedMatmulTest = MultiDeviceTest;
 TEST_F(OverlapDistributedMatmulTest, AG_matmul) {
   // Disable StreamParallelType pass temporarily as proper stream lowering gets
   // implemented
-  preseg_passes::OptimizationPassGuard<hir::StreamParallelType> guard(false);
+  hir_pass::OptimizationPassGuard<hir_pass::StreamParallelType> guard(false);
 
   constexpr int64_t M = 32768;
   constexpr int64_t K = 32768;
@@ -423,7 +423,7 @@ TEST_F(OverlapDistributedMatmulTest, AG_matmul) {
 
 TEST_F(OverlapDistributedMatmulTest, AG_linear) {
   // Disable StreamParallelType pass tempor
-  preseg_passes::OptimizationPassGuard<hir::StreamParallelType> guard(false);
+  hir_pass::OptimizationPassGuard<hir_pass::StreamParallelType> guard(false);
 
   constexpr int64_t M = 32768;
   constexpr int64_t K = 32768;
